The Music Update --Laid the groundwork for other modules in the future --Added the ability for users to play music in voice channels --Added the !join command to join a voice channel --Added the !leave command to leave a voice channel --Added the !play command to play a youtube link --Added the !stop command to stop playback --Removed the News feature in preparation for a total overall. We will soon be basing the feature in the twitter API instead
Known Issue:
 --!newevent does not support its optional parameters. A fix is in the works for this issue. Please use !editevent in the meantime
