+++ HavocBot/Classes/customColors.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace HavocBot
+{
+    /// <summary>
+    /// 
+    /// </summary>
+    public class customColors
+    {
+    }
+}
