+++ HavocBot/Resources/opus.dll.checksums
@@ -1,0 +1,3 @@
+SHA256 (opus.dll) = aef1cf46c9c6515d57a104dc945324bbe2411c383b6c71f35fb7715651ba47fa
+SHA384 (opus.dll) = e69a8772598748fd5a155a65004bd5233fac3ed7b533af6c951f147ed944eec0c5d2c6774e94c265e8863808f7b581fb
+SHA512 (opus.dll) = dd0c397e36dde48f628e511e76ee534efbc5ec8a181c4a2cf095f92b0f4222e3deba4702e931a8f54b0ef13d1052a956a8259ecc042ddaec882fa107c2dad704
