+++ HavocBot/bin/Release/netcoreapp3.0/publish/HavocBot.deps.json
@@ -30,7 +30,8 @@
           "Discord.Net.Rest": "2.1.1",
           "Discord.Net.WebSocket": "2.1.1",
           "Discord.Net.Webhook": "2.1.1",
-          "HtmlAgilityPack": "1.11.18",
+          "HtmlAgilityPack": "1.11.21",
+          "HtmlAgilityPack.CssSelectors": "1.0.2",
           "Microsoft.CodeAnalysis.FxCopAnalyzers": "2.9.8",
           "Newtonsoft.Json": "12.0.3",
           "dcsoup": "1.0.0",
@@ -1791,7 +1792,7 @@
           "lib/netstandard2.0/Discord.Net.WebSocket.dll": {}
         }
       },
-      "HtmlAgilityPack/1.11.18": {
+      "HtmlAgilityPack/1.11.21": {
         "dependencies": {
           "System.Net.Http": "4.3.2",
           "System.Xml.XPath": "4.3.0",
@@ -1800,6 +1801,14 @@
         },
         "compile": {
           "lib/netstandard2.0/HtmlAgilityPack.dll": {}
+        }
+      },
+      "HtmlAgilityPack.CssSelectors/1.0.2": {
+        "dependencies": {
+          "HtmlAgilityPack": "1.11.21"
+        },
+        "compile": {
+          "lib/net45/HtmlAgilityPack.CssSelectors.dll": {}
         }
       },
       "Microsoft.CodeAnalysis.FxCopAnalyzers/2.9.8": {
@@ -2772,7 +2781,8 @@
           "Discord.Net.Rest": "2.1.1",
           "Discord.Net.WebSocket": "2.1.1",
           "Discord.Net.Webhook": "2.1.1",
-          "HtmlAgilityPack": "1.11.18",
+          "HtmlAgilityPack": "1.11.21",
+          "HtmlAgilityPack.CssSelectors": "1.0.2",
           "Microsoft.CodeAnalysis.FxCopAnalyzers": "2.9.8",
           "Newtonsoft.Json": "12.0.3",
           "dcsoup": "1.0.0",
@@ -3141,7 +3151,7 @@
           }
         }
       },
-      "HtmlAgilityPack/1.11.18": {
+      "HtmlAgilityPack/1.11.21": {
         "dependencies": {
           "System.Net.Http": "4.3.2",
           "System.Xml.XPath": "4.3.0",
@@ -3150,8 +3160,19 @@
         },
         "runtime": {
           "lib/netstandard2.0/HtmlAgilityPack.dll": {
-            "assemblyVersion": "1.11.18.0",
-            "fileVersion": "1.11.18.0"
+            "assemblyVersion": "1.11.21.0",
+            "fileVersion": "1.11.21.0"
+          }
+        }
+      },
+      "HtmlAgilityPack.CssSelectors/1.0.2": {
+        "dependencies": {
+          "HtmlAgilityPack": "1.11.21"
+        },
+        "runtime": {
+          "lib/net45/HtmlAgilityPack.CssSelectors.dll": {
+            "assemblyVersion": "1.0.0.0",
+            "fileVersion": "1.0.0.0"
           }
         }
       },
@@ -4181,12 +4202,19 @@
       "path": "discord.net.websocket/2.1.1",
       "hashPath": "discord.net.websocket.2.1.1.nupkg.sha512"
     },
-    "HtmlAgilityPack/1.11.18": {
+    "HtmlAgilityPack/1.11.21": {
       "type": "package",
       "serviceable": true,
-      "sha512": "sha512-Wdfc2rEy0nr1peawYnOA8aufAo9Z5y16Oz4Rr+NmN8kMB0YBbHO4igYb+a0Px4LX5Ehx2u70dSnU6Be05Tna+w==",
-      "path": "htmlagilitypack/1.11.18",
-      "hashPath": "htmlagilitypack.1.11.18.nupkg.sha512"
+      "sha512": "sha512-WQdI36PMCl05+sQ0llvVW641YdCbwVCKvIkhukMpg/AvGECvPtry6O0Rt2eOvEe5QKlyj0c3Jv/RNaP2ZIoP3Q==",
+      "path": "htmlagilitypack/1.11.21",
+      "hashPath": "htmlagilitypack.1.11.21.nupkg.sha512"
+    },
+    "HtmlAgilityPack.CssSelectors/1.0.2": {
+      "type": "package",
+      "serviceable": true,
+      "sha512": "sha512-bB4FeVvcNROENPWQfFSmPQ6q99jGk8ru5QoBB3qnEeaP451fjnqQ0uPuMgmWHAOZIXUAoOQozv/5fEEVSDyIzA==",
+      "path": "htmlagilitypack.cssselectors/1.0.2",
+      "hashPath": "htmlagilitypack.cssselectors.1.0.2.nupkg.sha512"
     },
     "Microsoft.CodeAnalysis.FxCopAnalyzers/2.9.8": {
       "type": "package",
