+++ libsodium.dll.checksums
@@ -1,0 +1,3 @@
+SHA256 (libsodium.dll) = f6596837901f07475885970a394feba71a8dd4582ff678dc8ae6078b95dba903
+SHA384 (libsodium.dll) = 9f9bee759893083b0717be42efe9eb17c45f77854ead3c770e5de4ac4ad373d521224e64a82f41a68867dc6803e99168
+SHA512 (libsodium.dll) = 7dd28656ecee7e83ba4b8021c904264531ad0f66c6f76f5f2d110d17231a2c1685554d5690d9b70aff6de9fa6b28fa4749d2498faaf811991a811e606c6d0b9b
